+++ src/main/java/org/example/security/Principal.java
@@ -3,13 +3,15 @@ package org.example.security;
 import lombok.Data;
 import org.example.user.Role;
 
+import java.math.BigDecimal;
+
 @Data
 public class Principal {
 
-    private Long id;
+    private Integer id;
     private String email;
     private String username;
     private Role role;
-    private double balance;
+    private BigDecimal balance;
 
 }
