+++ src/main/java/org/example/ticket/Ticket.java
@@ -5,6 +5,7 @@ import lombok.Setter;
 import lombok.ToString;
 import org.example.ticket.StatusForTicket;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -15,7 +16,7 @@ public class Ticket {
     private Integer userId;
     private Integer outcomeId;
     private StatusForTicket status;
-    private double stake;
+    private BigDecimal stake;
     private String resultName;
     private String eventName;
     private LocalDateTime eventStartTime;
